backend/requirements.txt
supabase
python-dotenv
python-escpos